requirements.txt
dateutil
bs4
requests
pandas
lxml
html5lib
python-telegram-bot
PyMySQL
